# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Varkiv HTTP API
  summary: Stable, versioned contract for the Varkiv library and device ecosystem
  version: 0.1.0-preview.1
  description: |
    Versioned HTTP API for a self-hosted personal ROM catalog, frontend imports
    and exports, scoped device pairing, multi-file save streams, and automatic
    sync sessions. Every API response includes `X-Request-ID`,
    `X-Varkiv-API-Version`, and `Cache-Control: no-store`.

    Unless an operation explicitly declares `security: []`, deployments with an
    owner token require `Authorization: Bearer <token>`. Device tokens are opaque,
    independently revocable credentials whose effective scopes are enforced by
    the server. Signed content/session URLs are short-lived capabilities rather
    than bearer tokens.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
...
servers:
  - url: /api/v1
    description: Same-origin stable API entry point
security:
  - bearerAuth: []
...
paths:
...
  /web-emulation/saves/{token}:
    parameters:
      - {name: token, in: path, required: true, schema: {type: string}}
    get:
      operationId: getWebEmulationSavesByToken
      tags: [Saves]
      summary: Restore the current save for a signed browser-play session
      security: []
      responses:
        '200':
          description: Current core-isolated battery save
          headers:
            X-Varkiv-Revision: {schema: {type: string}, description: Revision to send as the base of the next upload}
          content: {application/octet-stream: {schema: {type: string, format: binary}}}
        '204': {description: The stream has no save revision yet}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
    post:
      operationId: postWebEmulationSavesByToken
      tags: [Saves]
      summary: Add an automatic core-isolated save revision
      description: The signed play capability selects Edition, exact EmulatorJS core, driver, and SaveStream. A stale base is preserved as a conflict instead of overwriting history.
      security: []
      parameters:
        - {name: X-Varkiv-Base-Revision, in: header, schema: {type: string}}
      requestBody:
        required: true
        content: {application/octet-stream: {schema: {type: string, format: binary, maxLength: 33554432}}}
      responses:
        '201': {description: Immutable save revision created or deduplicated, content: {application/json: {schema: {$ref: '#/components/schemas/SavePushResult'}}}}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
...
  /devices:
    get:
      operationId: getDevices
      tags: [Devices]
      summary: List registered devices
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated devices, content: {application/json: {schema: {$ref: '#/components/schemas/DeviceCollection'}}}}
    post:
      operationId: postDevices
      tags: [Devices]
      summary: Register a device
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DeviceMutation'}}}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '201': {description: Registered device, content: {application/json: {schema: {$ref: '#/components/schemas/Device'}}}}
...
  /pairing-codes/redeem:
    post:
      operationId: postPairingCodesRedeem
      tags: [Devices]
      summary: Redeem a pairing code and receive a client bearer token once
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PairingRedeemRequest'}
      responses:
        '201': {description: Paired device and one-time plaintext token, content: {application/json: {schema: {$ref: '#/components/schemas/PairingRedeemResponse'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '410': {$ref: '#/components/responses/Error'}
  /saves:
    get:
      operationId: getSaves
      tags: [Saves]
      summary: List immutable save revisions
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {name: edition_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated save revisions, content: {application/json: {schema: {$ref: '#/components/schemas/SaveRevisionCollection'}}}}
...
  /saves/{id}:
    get:
      operationId: getSavesById
      tags: [Saves]
      summary: Get immutable save revision metadata
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: Save revision metadata, content: {application/json: {schema: {$ref: '#/components/schemas/SaveRevision'}}}}
        '404': {$ref: '#/components/responses/Error'}
...
  /sync/manifest:
    get:
      operationId: getSyncManifest
      tags: [Sync]
      summary: Get the compact Edition and ROM identity manifest for device clients
      description: Clients match local ROMs to an Edition, then use its stable save_namespace when syncing emulator save files.
      parameters:
        - {$ref: '#/components/parameters/Locale'}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200':
          description: Device sync identity manifest
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SyncManifest'}
...
  /sync/inventory-matches:
    get:
      operationId: getSyncInventoryMatches
      tags: [Sync]
      summary: List ambiguous ROM identity matches that require an administrator decision
      description: Administrator-only. Responses expose catalog candidates and an opaque inventory record identifier, but never the client-local ROM path, filename, hash, serial, product code, title ID, or opaque client item identity. Paired device tokens are rejected.
      parameters:
        - {$ref: '#/components/parameters/Locale'}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
      responses:
        '200': {description: Paginated ambiguous inventory matches, content: {application/json: {schema: {$ref: '#/components/schemas/InventoryMatchReviewCollection'}}}}
        '403': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Owner token or a server-issued device token; device scopes are enforced per route.
  headers:
    RequestID:
      description: Caller-supplied valid request identity or a server-generated identity.
      schema: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._-]+$'}
    APIVersion:
      description: Stable API path version used for this response.
      schema: {type: string, const: v1}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string}, description: Opaque resource identifier; clients must not infer its shape or ordering.}
    Locale: {name: locale, in: query, schema: {type: string, enum: [zh-CN, zh-TW, ja, en]}, description: Preferred localized title. The resource default remains the fallback.}
...
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}, description: Maximum number of records returned in this page.}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}, description: Zero-based record offset in the endpoint's documented stable ordering.}
...
  responses:
...
    Error:
      description: Structured API error
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestID'}
        X-Varkiv-API-Version: {$ref: '#/components/headers/APIVersion'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          example:
            error:
              code: invalid_argument
              message: request payload is invalid
              request_id: 40e39da4b75c41c28cf40b9f8de46c48
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: {type: string, description: Stable machine-readable error code used for client branching.}
            message: {type: string, description: Human-readable diagnostic; clients must not parse this value.}
            request_id: {type: string, description: Correlation identity also returned in the X-Request-ID header.}
    Pagination:
      type: object
      additionalProperties: false
      description: Offset page metadata. Repeat the same filters and ordering while advancing offset by the number of returned records.
      required: [limit, offset, total]
      properties:
        limit: {type: integer, minimum: 1, maximum: 200}
        offset: {type: integer, minimum: 0}
        total: {type: integer, minimum: 0}
...
    DeviceMutation:
      type: object
      additionalProperties: false
      required: [name, os_family]
      properties:
        id: {type: string}
        name: {type: string, minLength: 1}
        device_profile_id: {type: string}
        os_family: {type: string, minLength: 1}
        distribution: {type: string}
        architecture: {type: string}
        agent_version: {type: string}
        status: {type: string, enum: [active, offline]}
        capabilities: {type: object, additionalProperties: {type: boolean}}
...
    SavePushResult:
      type: object
      additionalProperties: false
      required: [revision, created, conflict]
      properties:
        revision: {$ref: '#/components/schemas/SaveRevision'}
        created: {type: boolean}
        conflict: {type: boolean}
...
    Device:
      type: object
      required: [id, name, os_family, status, capabilities, created_at, updated_at, last_seen_at]
      properties:
        id: {type: string}
        name: {type: string}
        device_profile_id: {type: string}
        os_family: {type: string}
        distribution: {type: string}
        architecture: {type: string}
        agent_version: {type: string}
        status: {type: string, enum: [active, offline, revoked]}
        capabilities: {type: object, additionalProperties: {type: boolean}}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        last_seen_at: {type: string, format: date-time}
        revoked_at: {type: string, format: date-time}
...
    PairingRedeemRequest:
      type: object
      required: [code, device]
      properties:
        code: {type: string}
        device:
          type: object
          required: [name, os_family]
          properties:
            name: {type: string}
            device_profile_id: {type: string, description: Optional legacy assertion. New clients omit it; when present it must exactly match the administrator-bound profile and can never select or replace that profile.}
            os_family: {type: string}
            distribution: {type: string}
            architecture: {type: string}
            agent_version: {type: string}
            capabilities: {type: object, additionalProperties: {type: boolean}}
    PairingRedeemResponse:
      type: object
      required: [device, device_target, access_token, token_type, expires_at, scopes]
      properties:
        device: {$ref: '#/components/schemas/Device'}
        device_target: {type: string, description: Stable target of the enabled DeviceProfile selected by the administrator; empty only for legacy profile-less pairing.}
        access_token: {type: string, writeOnly: true, description: Returned once; the server persists only a SHA-256 hash.}
        token_type: {type: string, const: Bearer}
        expires_at: {type: string, format: date-time}
        scopes: {type: array, items: {type: string, enum: ['sync:read', 'sync:write', 'device:heartbeat']}}
...
    SaveFile:
      type: object
      required: [id, revision_id, logical_path, checksum, size]
      properties:
        id: {type: string}
        revision_id: {type: string}
        logical_path: {type: string, description: Portable relative path; server blob paths are never exposed.}
        checksum: {type: string, pattern: '^[a-f0-9]{64}$'}
        size: {type: integer, format: int64, minimum: 0}
        mtime_ns: {type: integer, format: int64}
        mode: {type: integer, format: int64}
    SaveRevision:
      type: object
      required: [id, stream_id, content_hash, total_size, file_count, status, files, device_id, created_at]
      properties:
        id: {type: string}
        stream_id: {type: string}
        device_id: {type: string}
        parent_revision_id: {type: string}
        content_hash: {type: string, pattern: '^[a-f0-9]{64}$'}
        total_size: {type: integer, format: int64, minimum: 0}
        file_count: {type: integer, minimum: 1, maximum: 4096}
        status: {type: string, enum: [current, superseded, conflict, quarantined]}
        files: {type: array, minItems: 1, items: {$ref: '#/components/schemas/SaveFile'}}
        created_at: {type: string, format: date-time}
...
    InventoryMatchCandidate:
      type: object
      required: [edition_id, game_id, game_title, edition_title, edition_type, platform_id]
      properties:
        edition_id: {type: string}
        game_id: {type: string}
        game_title: {type: string}
        edition_title: {type: string}
        edition_type: {type: string}
        platform_id: {type: string}
    InventoryMatchReview:
      type: object
      description: Privacy-minimized administrator review; local ROM names, paths, identifiers, and hashes are deliberately absent.
      required: [session_id, inventory_item_id, device_name, platform_id, match_method, size, candidates]
      properties:
        session_id: {type: string}
        inventory_item_id: {type: string, description: Opaque server-side inventory row identifier.}
        device_name: {type: string}
        platform_id: {type: string}
        match_method: {type: string, enum: [serial, product_code, title_id]}
        size: {type: integer, format: int64, minimum: 0}
        candidates: {type: array, minItems: 2, items: {$ref: '#/components/schemas/InventoryMatchCandidate'}}
...
    SyncArtifact:
      type: object
      required: [id, path, role, size, missing]
      properties:
        id: {type: string, format: uuid}
        path: {type: string}
        role: {type: string, enum: [rom, disc, executable]}
        disc_index: {type: integer, minimum: 0, maximum: 64}
        size: {type: integer, format: int64}
        sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        missing: {type: boolean}
    SyncEdition:
      type: object
      required: [game_id, game_title, platform, edition_id, edition_title, save_namespace, artifacts, revision_count]
      properties:
        game_id: {type: string, format: uuid}
        game_title: {type: string}
        platform: {type: string}
        edition_id: {type: string, format: uuid}
        edition_title: {type: string}
        save_namespace: {type: string}
        serial: {type: string}
        product_code: {type: string}
        title_id: {type: string}
        artifacts: {type: array, items: {$ref: '#/components/schemas/SyncArtifact'}}
        revision_count: {type: integer, minimum: 0}
        latest_revision: {type: object}
    SyncManifest:
      type: object
      required: [api_version, matching_order, editions]
      properties:
        api_version: {type: string}
        matching_order: {type: array, items: {type: string, enum: [sha256, serial, product_code, title_id]}}
        editions: {type: array, items: {$ref: '#/components/schemas/SyncEdition'}}
...
    InventoryMatchReviewCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/InventoryMatchReview'}}}}
...
    DeviceCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Device'}}}}
    SaveRevisionCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/SaveRevision'}}}}
...
    PaginatedBase:
      type: object
      required: [data, pagination]
      properties:
        data: {type: array, items: {type: object}}
        pagination: {$ref: '#/components/schemas/Pagination'}